Update search_algorithms.py
I had to change some of the code again because I noticed that it returned that only the indices and not the values of the PSUs were sorted and then the highest was returned. But we want to return the PSU with the highest value. Therefore, I added the lines from beam search, so that we have the value and the original index. Now if we sort with sorted(), the function looks at the value of the PSU and not its original index. Thereby, best will return the index of the PSU with the highest value.